refactoring: move apiClient to api/
