+++ webapp/src/api/index.js
@@ -1,1 +1,2 @@
 export { default as tryParseCommand } from './Parser'
+export { default as apiClient } from './ApiClient'
